+++ sniffer/utils/funcutils.py
@@ -2,6 +2,7 @@ import sys
 import rich
 import subprocess
 from rich.panel import Panel
+
 from . import constants