+++ sniffer/network/layers.py
@@ -1,11 +1,21 @@
 import struct
 import ipaddress
-from ..exceptions.network import UnsupportedVersionException, \
-    UninterestingPacketException
+
+from ..exceptions.network import UnsupportedVersionException
 
 
 class Layer3:
-    """Parse the layer 3 data of a packet."""
+    """Parse the layer 3 data of a packet.
+
+    Attributes
+    ----------
+    version : int
+        The IP protocol version.
+    source_ip : IPv4Address
+        The source IP address of the packet.
+    dest_ip : IPv4Address
+        The destination IP address of the packet.
+    """
     IP_VERSION_4 = 0x45
 
     def __init__(self, packet_bytes: bytes):
@@ -22,7 +32,15 @@ class Layer3:
 
 
 class Layer4:
-    """Parse the layer 4 data of a packet."""
+    """Parse the layer 4 data of a packet.
+
+    Attributes
+    ----------
+    source_port : int
+        The source port.
+    dest_port : int
+        The destination port.
+    """
     def __init__(self, packet_bytes: bytes):
         # Ignoring 28 bytes
         tcp_header = struct.unpack('!HH28s', packet_bytes[34:66])
@@ -32,6 +50,12 @@ class Layer4:
 
 
 class Layer7:
-    """Parse the layer 7 data of a packet (that's the actual data)."""
+    """Parse the layer 7 data of a packet (that's the actual data).
+
+    Attributes
+    ----------
+    data : bytes
+        The data of the packet, without headers.
+    """
     def __init__(self, packet_bytes: bytes):
         self.data = packet_bytes[66:]