+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.2] - 2021-12-31
+### Added
+- Add missing docstrings
+
+### Changed
+- Put the text in the side panel into 2 separate panels
+- Fixed incomplete docstrings
+
 ## [0.5.1] - 2021-12-30
 ### Added
 - Implement `-f/--filter` option with the following filters:
@@ -124,7 +132,7 @@ which re-arranges the layout and adds coloring 🌠
 the version number and Git commit
 
 
-[Unreleased]: https://github.com/reloadedd/gui-http-sniffer/compare/v0.5.1...HEAD
+[Unreleased]: https://github.com/reloadedd/gui-http-sniffer/compare/v0.5.2...HEAD
 [0.1.0]: https://github.com/reloadedd/gui-http-sniffer/releases/tag/v0.1.0
 [0.1.1]: https://github.com/reloadedd/gui-http-sniffer/releases/tag/v0.1.1
 [0.2.0]: https://github.com/reloadedd/gui-http-sniffer/releases/tag/v0.2.0
@@ -135,3 +143,4 @@ the version number and Git commit
 [0.4.0]: https://github.com/reloadedd/gui-http-sniffer/releases/tag/v0.4.0
 [0.5.0]: https://github.com/reloadedd/gui-http-sniffer/releases/tag/v0.5.0
 [0.5.1]: https://github.com/reloadedd/gui-http-sniffer/releases/tag/v0.5.1
+[0.5.2]: https://github.com/reloadedd/gui-http-sniffer/releases/tag/v0.5.2